bordel

+++ C_export/include/k2c_activations.c
@@ -19,7 +19,7 @@ https://github.com/f0uriest/keras2c
  * :param x: array of input values. Gets overwritten by output.
  * :param size: length of input array.
  */
-void k2c_linear_func(float * x, const size_t size) {
+void k2c_linear_func(double * x, const size_t size) {
 
 }
 k2c_activationType * k2c_linear = k2c_linear_func;
@@ -32,7 +32,7 @@ k2c_activationType * k2c_linear = k2c_linear_func;
  * :param x: array of input values. Gets overwritten by output.
  * :param size: length of input array.
  */
-void k2c_exponential_func(float * x, const size_t size) {
+void k2c_exponential_func(double * x, const size_t size) {
 
     for (size_t i=0; i<size; ++i) {
         x[i] = expf(x[i]);
@@ -48,7 +48,7 @@ k2c_activationType * k2c_exponential = k2c_exponential_func;
  * :param x: array of input values. Gets overwritten by output.
  * :param size: length of input array.
  */
-void k2c_relu_func(float * x, const size_t size) {
+void k2c_relu_func(double * x, const size_t size) {
 
     for (size_t i=0; i < size; ++i) {
         if (x[i] <= 0.0f) {
@@ -68,7 +68,7 @@ k2c_activationType * k2c_relu = k2c_relu_func;
  * :param x: array of input values. Gets overwritten by output.
  * :param size: length of input array.
  */
-void k2c_hard_sigmoid_func(float * x, const size_t size) {
+void k2c_hard_sigmoid_func(double * x, const size_t size) {
 
     for (size_t i=0; i < size; ++i) {
         if (x[i] <= -2.5f) {
@@ -92,7 +92,7 @@ k2c_activationType * k2c_hard_sigmoid = k2c_hard_sigmoid_func;
  * :param x: array of input values. Gets overwritten by output.
  * :param size: length of input array.
  */
-void k2c_tanh_func(float * x, const size_t size) {
+void k2c_tanh_func(double * x, const size_t size) {
 
     for (size_t i=0; i<size; ++i) {
         x[i] = tanhf(x[i]);
@@ -108,7 +108,7 @@ k2c_activationType * k2c_tanh = k2c_tanh_func;
  * :param x: array of input values. Gets overwritten by output.
  * :param size: length of input array.
  */
-void k2c_sigmoid_func(float * x, const size_t size) {
+void k2c_sigmoid_func(double * x, const size_t size) {
 
     for (size_t i=0; i < size; ++i) {
         x[i] = 1/(1+expf(-x[i]));
@@ -125,10 +125,10 @@ k2c_activationType * k2c_sigmoid = k2c_sigmoid_func;
  * :param x: array of input values. Gets overwritten by output.
  * :param size: length of input array.
  */
-void k2c_softmax_func(float * x, const size_t size) {
+void k2c_softmax_func(double * x, const size_t size) {
 
-    float xmax = x[0];
-    float sum = 0;
+    double xmax = x[0];
+    double sum = 0;
     for (size_t i=0; i < size; ++i) {
         if (x[i]>xmax) {
             xmax = x[i];
@@ -158,7 +158,7 @@ k2c_activationType * k2c_softmax = k2c_softmax_func;
  * :param x: array of input values. Gets overwritten by output.
  * :param size: length of input array.
  */
-void k2c_softplus_func(float * x, const size_t size) {
+void k2c_softplus_func(double * x, const size_t size) {
 
     for (size_t i=0; i < size; ++i) {
         x[i] = log1pf(expf(x[i]));
@@ -174,7 +174,7 @@ k2c_activationType * k2c_softplus = k2c_softplus_func;
  * :param x: array of input values. Gets overwritten by output.
  * :param size: length of input array.
  */
-void k2c_softsign_func(float * x, const size_t size) {
+void k2c_softsign_func(double * x, const size_t size) {
 
     for (size_t i=0; i < size; ++i) {
         x[i] = x[i]/(1.0f + fabsf(x[i]));
@@ -193,7 +193,7 @@ k2c_activationType * k2c_softsign = k2c_softsign_func;
  * :param size: length of input array.
  * :param alpha: slope of negative portion of activation curve.
  */
-void k2c_LeakyReLU(float * x, const size_t size, const float alpha) {
+void k2c_LeakyReLU(double * x, const size_t size, const double alpha) {
 
     for (size_t i=0; i<size; ++i) {
         if (x[i]<0) {
@@ -214,7 +214,7 @@ void k2c_LeakyReLU(float * x, const size_t size, const float alpha) {
  * :param size: length of input array.
  * :param alpha: slope of negative portion of activation curve for each unit.
  */
-void k2c_PReLU(float * x, const size_t size, const float * alpha) {
+void k2c_PReLU(double * x, const size_t size, const double * alpha) {
 
     for (size_t i=0; i<size; ++i) {
         if (x[i]<0.0f) {
@@ -233,7 +233,7 @@ void k2c_PReLU(float * x, const size_t size, const float * alpha) {
  * :param size: length of input array.
  * :param alpha: slope of negative portion of activation curve.
  */
-void k2c_ELU(float * x, const size_t size, const float alpha) {
+void k2c_ELU(double * x, const size_t size, const double alpha) {
 
     for (size_t i=0; i < size; ++i) {
         if (x[i] <= 0.0f) {
@@ -252,7 +252,7 @@ void k2c_ELU(float * x, const size_t size, const float alpha) {
  * :param size: length of input array.
  * :param theta: threshold for activation.
  */
-void k2c_ThresholdedReLU(float * x, const size_t size, const float theta) {
+void k2c_ThresholdedReLU(double * x, const size_t size, const double theta) {
 
     for (size_t i=0; i<size; ++i) {
         if (x[i]<= theta) {
@@ -273,8 +273,8 @@ void k2c_ThresholdedReLU(float * x, const size_t size, const float theta) {
  * :param alpha: slope of negative portion of activation curve.
  * :param theta: threshold for activation.
  */
-void k2c_ReLU(float * x, const size_t size, const float max_value,
-              const float alpha, const float theta) {
+void k2c_ReLU(double * x, const size_t size, const double max_value,
+              const double alpha, const double theta) {
 
     for (size_t i=0; i<size; ++i) {
         if (x[i] >= max_value) {
