+++ C_export/include/k2c_recurrent_layers.c
@@ -24,8 +24,8 @@ https://github.com/f0uriest/keras2c
  * :param recurrent_activation: activation function to apply to internal state.
  * :param output_activation: activation function to apply to output.
  */
-void k2c_lstmcell(float * state, const float * input, const k2c_tensor* kernel,
-                  const k2c_tensor* recurrent_kernel, const k2c_tensor* bias, float * fwork,
+void k2c_lstmcell(double * state, const double * input, const k2c_tensor* kernel,
+                  const k2c_tensor* recurrent_kernel, const k2c_tensor* bias, double * fwork,
                   k2c_activationType *recurrent_activation,
                   k2c_activationType *output_activation) {
 
@@ -33,29 +33,29 @@ void k2c_lstmcell(float * state, const float * input, const k2c_tensor* kernel,
     const size_t units = recurrent_kernel->shape[1];
     const size_t in_width = kernel->shape[0]/4;
 
-    float *h_tm1 = &state[0];  // previous memory state
-    float *c_tm1 = &state[units];  // previous carry state
+    double *h_tm1 = &state[0];  // previous memory state
+    double *c_tm1 = &state[units];  // previous carry state
     const size_t outrows = 1;
-    const float * const Wi = &kernel->array[0];
-    const float * const Wf = &kernel->array[in_width*units];
-    const float * const Wc = &kernel->array[2*in_width*units];
-    const float * const Wo = &kernel->array[3*in_width*units];
-    const float * const Ui = &recurrent_kernel->array[0];
-    const float * const Uf = &recurrent_kernel->array[units*units];
-    const float * const Uc = &recurrent_kernel->array[2*units*units];
-    const float * const Uo = &recurrent_kernel->array[3*units*units];
-    const float * const bi = &bias->array[0];
-    const float * const bf = &bias->array[units];
-    const float * const bc = &bias->array[2*units];
-    const float * const bo = &bias->array[3*units];
-    float *xi = &fwork[0];
-    float *xf = &fwork[units];
-    float *xc = &fwork[2*units];
-    float *xo = &fwork[3*units];
-    float *yi = &fwork[4*units];
-    float *yf = &fwork[5*units];
-    float *yc = &fwork[6*units];
-    float *yo = &fwork[7*units];
+    const double * const Wi = &kernel->array[0];
+    const double * const Wf = &kernel->array[in_width*units];
+    const double * const Wc = &kernel->array[2*in_width*units];
+    const double * const Wo = &kernel->array[3*in_width*units];
+    const double * const Ui = &recurrent_kernel->array[0];
+    const double * const Uf = &recurrent_kernel->array[units*units];
+    const double * const Uc = &recurrent_kernel->array[2*units*units];
+    const double * const Uo = &recurrent_kernel->array[3*units*units];
+    const double * const bi = &bias->array[0];
+    const double * const bf = &bias->array[units];
+    const double * const bc = &bias->array[2*units];
+    const double * const bo = &bias->array[3*units];
+    double *xi = &fwork[0];
+    double *xf = &fwork[units];
+    double *xc = &fwork[2*units];
+    double *xo = &fwork[3*units];
+    double *yi = &fwork[4*units];
+    double *yf = &fwork[5*units];
+    double *yc = &fwork[6*units];
+    double *yo = &fwork[7*units];
 
 
     //xi = input*Wi + bi;
@@ -117,9 +117,9 @@ void k2c_lstmcell(float * state, const float * input, const k2c_tensor* kernel,
  * :param recurrent_activation: activation function to apply to internal state.
  * :param output_activation: activation function to apply to output.
  */
-void k2c_lstm(k2c_tensor* output, const k2c_tensor* input, float * state,
+void k2c_lstm(k2c_tensor* output, const k2c_tensor* input, double * state,
               const k2c_tensor* kernel, const k2c_tensor* recurrent_kernel,
-              const k2c_tensor* bias, float * fwork, const int go_backwards,
+              const k2c_tensor* bias, double * fwork, const int go_backwards,
               const int return_sequences, k2c_activationType *recurrent_activation,
               k2c_activationType *output_activation) {
 
@@ -169,16 +169,16 @@ void k2c_lstm(k2c_tensor* output, const k2c_tensor* input, float * state,
  * :param fwork: array[2*units] working storage.
  * :param output_activation: activation function to apply to output.
  */
-void k2c_simpleRNNcell(float * state, const float * input, const k2c_tensor* kernel,
+void k2c_simpleRNNcell(double * state, const double * input, const k2c_tensor* kernel,
                        const k2c_tensor* recurrent_kernel, const k2c_tensor* bias,
-                       float * fwork, k2c_activationType *output_activation) {
+                       double * fwork, k2c_activationType *output_activation) {
 
     const size_t units = recurrent_kernel->shape[1];
     const size_t in_width = kernel->shape[0];
 
     const size_t outrows = 1;
-    float *h1 = &fwork[0];
-    float *h2 = &fwork[units];
+    double *h1 = &fwork[0];
+    double *h2 = &fwork[units];
     // h1 = input*kernel+bias
     k2c_affine_matmul(h1,input,kernel->array,bias->array,outrows,units,in_width);
 
@@ -207,9 +207,9 @@ void k2c_simpleRNNcell(float * state, const float * input, const k2c_tensor* ker
  * :param return_sequences: whether to return the last output in the output sequence (0), or the full sequence (1).
  * :param output_activation: activation function to apply to output.
  */
-void k2c_simpleRNN(k2c_tensor* output, const k2c_tensor* input, float * state,
+void k2c_simpleRNN(k2c_tensor* output, const k2c_tensor* input, double * state,
                    const k2c_tensor* kernel, const k2c_tensor* recurrent_kernel,
-                   const k2c_tensor* bias, float * fwork, const int go_backwards,
+                   const k2c_tensor* bias, double * fwork, const int go_backwards,
                    const int return_sequences, k2c_activationType *output_activation) {
 
     const size_t in_width = input->shape[1];
@@ -260,34 +260,34 @@ void k2c_simpleRNN(k2c_tensor* output, const k2c_tensor* input, float * state,
  * :param recurrent_activation: activation function to apply to internal state.
  * :param output_activation: activation function to apply to output.
  */
-void k2c_grucell(float * state, const float * input, const k2c_tensor* kernel,
-                 const k2c_tensor* recurrent_kernel, const k2c_tensor* bias, float * fwork,
+void k2c_grucell(double * state, const double * input, const k2c_tensor* kernel,
+                 const k2c_tensor* recurrent_kernel, const k2c_tensor* bias, double * fwork,
                  const int reset_after, k2c_activationType *recurrent_activation,
                  k2c_activationType *output_activation) {
 
     const size_t units = recurrent_kernel->shape[1];
     const size_t in_width = kernel->shape[0]/3;
 
-    float *h_tm1 = &state[0];
+    double *h_tm1 = &state[0];
     const size_t outrows = 1;
-    const float * const Wz = &kernel->array[0];
-    const float * const Wr = &kernel->array[in_width*units];
-    const float * const Wh = &kernel->array[2*in_width*units];
-    const float * const Uz = &recurrent_kernel->array[0];
-    const float * const Ur = &recurrent_kernel->array[units*units];
-    const float * const Uh = &recurrent_kernel->array[2*units*units];
-    const float * const bz = &bias->array[0];
-    const float * const br = &bias->array[units];
-    const float * const bh = &bias->array[2*units];
-    const float * const rbz = &bias->array[3*units];
-    const float * const rbr = &bias->array[4*units];
-    const float * const rbh = &bias->array[5*units];
-    float *xz = &fwork[0];
-    float *xr = &fwork[units];
-    float *xh = &fwork[2*units];
-    float *yz = &fwork[3*units];
-    float *yr = &fwork[4*units];
-    float *yh = &fwork[5*units];
+    const double * const Wz = &kernel->array[0];
+    const double * const Wr = &kernel->array[in_width*units];
+    const double * const Wh = &kernel->array[2*in_width*units];
+    const double * const Uz = &recurrent_kernel->array[0];
+    const double * const Ur = &recurrent_kernel->array[units*units];
+    const double * const Uh = &recurrent_kernel->array[2*units*units];
+    const double * const bz = &bias->array[0];
+    const double * const br = &bias->array[units];
+    const double * const bh = &bias->array[2*units];
+    const double * const rbz = &bias->array[3*units];
+    const double * const rbr = &bias->array[4*units];
+    const double * const rbh = &bias->array[5*units];
+    double *xz = &fwork[0];
+    double *xr = &fwork[units];
+    double *xh = &fwork[2*units];
+    double *yz = &fwork[3*units];
+    double *yr = &fwork[4*units];
+    double *yh = &fwork[5*units];
 
     //     x_z = input*kernel_z + input_bias_z
     k2c_affine_matmul(xz, input, Wz, bz, outrows, units, in_width);
@@ -359,9 +359,9 @@ void k2c_grucell(float * state, const float * input, const k2c_tensor* kernel,
  * :param recurrent_activation: activation function to apply to internal state.
  * :param output_activation: activation function to apply to output.
  */
-void k2c_gru(k2c_tensor* output, const k2c_tensor* input, float * state,
+void k2c_gru(k2c_tensor* output, const k2c_tensor* input, double * state,
              const k2c_tensor* kernel, const k2c_tensor* recurrent_kernel,
-             const k2c_tensor* bias, float * fwork, const int reset_after,
+             const k2c_tensor* bias, double * fwork, const int reset_after,
              const int go_backwards, const int return_sequences,
              k2c_activationType *recurrent_activation,
              k2c_activationType *output_activation) {
